feat : Add approve preview on overview page
